Working points table

+++ prototype_1/lib/screens/guard_duty_tracker_screen.dart/add_new_duty_screen.dart
@@ -103,7 +103,7 @@ class _AddNewDutyScreenState extends State<AddNewDutyScreen> {
         .get()
         .then((value) {
           var data = value.data();
-          currentPoints = data!['points'];
+          currentPoints = data!['points'].toDouble();
         });
 
     await FirebaseFirestore.instance.collection('Users').doc(name).set({
